refactor(registry): revise parse_info for updated YAML spec format
- Replace top-level variable_dimension/input_dimension/output_dimension
  with nested dimensions section (input: int|"variable", output: int)
- Move default_input and default_parameters to top level
- Switch callable detection from "()" suffix to factory-key convention
- Update parameter sets structure (keyword block, reserved description)
- Update fixture YAMLs and tests to match

Closes: #474
Ref: #452
