+++ src/uqtestfuns/core/registry/__init__.py
@@ -5,5 +5,5 @@ Registry sub-package of UQTestFuns.
 from .registry import DEFAULT_ROOT, PKG_ROOT, Registry
 
 # Singleton registry
-_registry = Registry()
-_registry.scan(DEFAULT_ROOT, PKG_ROOT)
+_registry = Registry(PKG_ROOT)
+_registry.scan(DEFAULT_ROOT)
